CORTX:33252: INitial commit
